# inventory-service/src/main/resources/openapi/inventory-service.yaml
openapi: 3.0.3
info:
  title: Shipment API
  version: v1

tags:
  - name: Auth
    description: Endpoints for registration and login
  - name: Shipment
    description: Endpoints for shipment manipulation

paths:
  /public/auth/register:
    post:
      tags: [ Auth ]
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /public/auth/login:
    post:
      tags: [ Auth ]
      summary: Login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/v1/shipment:
    post:
      tags: [ Shipment ]
      summary: register shipment
      operationId: createShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShipmentRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
  /api/v1/shipment/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      tags: [ Shipment ]
      summary: get shipment
      operationId: getShipment
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
  /api/v1/shipmentUnit:
    post:
      tags: [ Shipment ]
      summary: register shipment unit
      operationId: createShipmentUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShipmentUnitRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentUnitResponse'
  /api/v1/shipmentUnit/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      tags: [ Shipment ]
      summary: get shipment unit
      operationId: getShipmentUnit
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentUnitResponse'

components:
  schemas:
    RegisterRequest:
      type: object
      required: [ username, password ]
      properties:
        username: { type: string }
        password: { type: string }

    LoginRequest:
      type: object
      required: [ username, password ]
      properties:
        username: { type: string }
        password: { type: string }

    AuthResponse:
      type: object
      properties:
        token: { type: string }

    CreateShipmentRequest:
      type: object
      required: [ supplierCode, productCode ]
      properties:
        supplierCode: { type: string }
        externalShipmentId: { type: string }
        productCode: { type: string }
    ShipmentResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        supplierCode: { type: string }
        externalShipmentId: { type: string }
        productCode: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateShipmentUnitRequest:
      type: object
      required: [ shipmentId, amount, lengthCm, widthCm, heightCm ]
      properties:
        shipmentId: { type: integer, format: int64 }
        amount: { type: integer, format: int64 }
        lengthCm: { type: integer, format: int64 }
        widthCm: { type: integer, format: int64 }
        heightCm: { type: integer, format: int64 }
    ShipmentUnitResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        shipmentId: { type: integer, format: int64 }
        shipmentStatus: { type: string }
        locationCode: { type: string }
        amount: { type: integer, format: int64 }
        lengthCm: { type: integer, format: int64 }
        widthCm: { type: integer, format: int64 }
        heightCm: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }